Fix an issue in KWS model and reference data
It was found that the NN model is not fully matching TensorFlow Lite Micro due to limitations in a conversion script. This patch corrects the model (ee_nn_tables.c) to fully match reference KWS model, and the reference data (kws_expected.c) is also updated accordingly. The patching was agreed by AudioMark workgroup members on Sept-5-2023.